1. [x]获取某签到活动内容详情 这个是pc管理端在慕课活动中，查看某签到活动内容的详情时，所用到的。  2. [x]增改继教签到活动内容  3.
 4.
[x]分页获取某文件夹下的签到活动
 5.
[x]获取某签到活动详情
 6.
[x]增改继教签到活动
 7.
[x]发布继教签到活动
 8.
[x]分页获取某签到的签到情况
 9.
[x]获取某人某签到的签到记录

+++ ScientificResearch/Models/Table/继教操作考试参与情况.cs
@@ -51,7 +51,7 @@ namespace ScientificResearch.Models
 		/// 
 		/// </summary>
         [Required(ErrorMessage ="请提供[得分]")]
-		public int 得分 { get; set; }
+		public decimal 得分 { get; set; }
         /// <summary>
 		/// 
 		/// </summary>
